+++ changelog.md
@@ -1,5 +1,12 @@
 # PHPMailer-AMP Change Log
 
+### Version 1.1.0
+* Content attachment with AMP support
+* Return back PHP <7.0 support
+
+#### Version 1.0.1
+* Rename namespaces
+
 ## Version 1.0.0
 * `AMPBody` property support
 * `isAMP()` function support